Implement the first iteration of CLI arg parsing
Essentially this commit sets the foundation for being able to parse a
config file that contains the information needed to run the collection
of servers that make up this project.

It is expected that more fields will need to be added but the included
fields are sufficient for a first attempt.