Move all defined types into their own file
We're moving types into their own file as they're going to need to be
used project wide. It's better to only have to import the types rather
than actual implementation details. This mitigates the chance of any
naming collisions.

In addition, this commit further expands on being able to read the
program configuration from disk and puts some placeholder instructions
for usage and installation.