Implement a minimal set of the commands for SMTP
A lot of work needs to be done to clean up the code especially with
regards to exception handling. Further work should look into using
the async library or slave thread library so that exception handlers
are installed top-level. In addition this would give us better semantics
about how we have the main thread block for it's children.